+++ metrics.ini
@@ -1154,7 +1154,6 @@ CursorTweenSeconds=0.15
 AllowRepeatingChangeValueInput=1
 ShowUnderlines=1
 CapitalizeAllOptionNames=0
-NumRowsShown=@gimmick.s.ScreenPlayerOptions.lines.NumRowsShown()
 NumIconParts=0
 NumCursorParts=0
 NumPreviewParts=0
@@ -1235,9 +1234,8 @@ Line1=@gimmick.s.ScreenPlayerOptions.lines.Line1()
 
 NumRowsShown=@gimmick.s.ScreenPlayerOptions.lines.NumRowsShown()
 
-FrameOnCommand=%function(s) gimmick.s.ScreenPlayerOptions.Options.Init(s) end
+FrameOnCommand=%function(s) gimmick.s.ScreenPlayerOptions.options.Init(s) end
 
-TimerSeconds=-1
 NextScreen=ScreenBranchStage
 PrevScreen=@SongSelectionScreen()
 
@@ -1253,22 +1251,9 @@ NextScreen=@gimmick.s.ScreenOptionsMenu.NextScreen()
 
 LineNames=@gimmick.s.ScreenOptionsMenu.lines.LineNames()
 Line1=@gimmick.s.ScreenOptionsMenu.lines.Line1()
-
-AllowRepeatingChangeValueInput=1
-ShowUnderlines=0
-CapitalizeAllOptionNames=0
 NumRowsShown=@gimmick.s.ScreenOptionsMenu.lines.NumRowsShown()
-NumIconParts=0
-NumCursorParts=0
-NumPreviewParts=0
 
-ShowScroller=1
-WrapScroller=0
-ShowScrollBar=0
-
-HelpOnCommand=hidden,1
-ExplanationTogetherOnCommand=hidden,1
-LineHighlightOnCommand=hidden,1
+FrameOnCommand=%function(s) gimmick.s.ScreenOptionsMenu.options.Init(s) end
 
 [ScreenEditMenu]
 InitCommand=%function(s) gimmick.s.ScreenEditMenu.Init(s) end
